refactor: 페이지 전환 에니메이션 효과 customHook으로 이동
